requirements.txt
fastapi>=0.116.0
uvicorn[standard]==0.44.0
firebase-admin>=6.0.0
playwright==1.47.0
anthropic==0.34.2
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
geopy==2.4.1
python-multipart==0.0.9
